Refactored no_delete.
