+++ safedelete/tests/test_hard_delete.py
@@ -1,4 +1,4 @@
-from . import SafeDeleteTestCase
+from .testcase import SafeDeleteTestCase
 from ..config import HARD_DELETE
 from ..models import SafeDeleteMixin
 
@@ -15,9 +15,3 @@ class SoftDeleteTestCase(SafeDeleteTestCase):
     def test_harddelete(self):
         """Deleting a model with the soft delete policy should only mask it, not delete it."""
         self.assertHardDelete(self.instance)
-
-    def test_harddelete_force(self):
-        self.assertHardDelete(self.instance, force=True)
-
-    def test_softdelete_force(self):
-        self.assertSoftDelete(self.instance, force=True)
